Add default Q learning to agent
Add code to create the Q dictionary for the state/action Q values and
add code to update these values when an action is taken. Use these Q
values to choose an action with probablility (1 - epsilon) instead of a
random action, so that we exploit our learning and decrement epsilon for
each new trial.