incorporate review feedback
Several changes based on feedback, the main one being to make a random
selection among several possible actions that have the same, max, Q
value. Previously I just chose the first one, which tended to be the
"None" action and led to an erroneously high safety score by effectively
injecting domain knowledge that "None" is a safe action when you are not
sure what to do. This meant I had to re-do most of the optimisation and
testing.
Other changes were improvments to answers to questions 2, 7 and 9